update to support multiple collectors for each location.
This quite hacky, but works.
The list of active clients connected is maintained in Redis.
When a client is added/removed from a location all clients
at that location are refreshed.

Additionally, each collector-ctrl node sends heartbeat
messages via rabbitmq. When a node goes offline the other active
nodes will remove the stale client information in redis and
initiate a refresh.